Remove the manifest php55 and replace with more flexible apt
The php55 manifest is a hacked together version of what the apt package is.
The apt package adds flexibility and allows you to define multiple apt repositories
to add simultaneously.

Fix for apt manifest multiple repositories

Add the $name variable to the exec. Puppet requires unique names for each class.
This allows you to add multiple repos to be added sequentially.